+++ odoo/local-src/mtsmte_stock/tests/__init__.py
@@ -3,3 +3,4 @@
 
 
 from . import test_stock_picking
+# from . import test_stock_scraping